Make suggested changes from code review
