+++ crates/fuel-core/src/query/tx.rs
@@ -1,3 +1,9 @@
+use crate::fuel_core_graphql_api::ports::{
+    DatabaseBlocks,
+    OffChainDatabase,
+    OnChainDatabase,
+};
+
 use fuel_core_storage::{
     iter::{
         BoxedIter,
@@ -17,12 +23,6 @@ use fuel_core_types::{
     },
     fuel_types::Address,
     services::txpool::TransactionStatus,
-};
-
-use crate::fuel_core_graphql_api::ports::{
-    DatabaseBlocks,
-    OffChainDatabase,
-    OnChainDatabase,
 };
 
 pub trait SimpleTransactionData: Send + Sync {
