+++ crates/fuel-core/src/service/adapters/gas_price_adapters/tests.rs
